adjusted for array jobs
git-svn-id: file:///srv/local/data/cleanRepo/trunk@22148 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/ikaddoura/src/main/java/playground/ikaddoura/busCorridorPaper/busCorridorWelfareAnalysis/OptSettingsReader.java
@@ -96,7 +96,7 @@ public class OptSettingsReader {
 	                		this.optSettings.setIncrCapacity(Integer.parseInt(parts[1]));
 	                	
 	                	} else if (parts[0].equals("randomSeed")){
-	                		this.optSettings.setRandomSeed(Integer.parseInt(parts[1]));
+	                		this.optSettings.setRandomSeed(Long.parseLong(parts[1]));
 	                		
 	            		} else {
 	            			throw new RuntimeException(parts[0] +" is an unknown parameter in the optimization settings file. Aborting...");
